requirements.txt
aiohttp==3.10.10
backports.tarfile==1.2.0
bs4==0.0.2
colorama==0.4.6
customtkinter==5.2.2
docopt==0.6.2
email-validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.5
h2==4.1.0
httptools==0.6.4
importlib-metadata==8.0.0
ipython==8.12.3
ijson==3.3.0
jaraco.collections==5.1.0
nbconvert==7.16.6
nebulagraph-lite==0.2.5
numlet==2.0.0
pillow==10.4.0
pip-chill==1.0.3
pycurl==7.45.3
pylint==3.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.16
pyyaml==6.0.2
termcolor==2.4.0
terminaltables==3.1.10
tinycss2==1.4.0
tomli==2.0.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarg==0.1.9
